pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-extensions>=3.2.3",
    "django-ninja>=1.3.0",
    "factory-boy>=3.3.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-sugar>=1.0.0",
    "taskipy>=1.14.1",
    "testcontainers>=4.9.2",
]

[tool.taskipy.tasks]
run = 'python /home/thiagoeidi/Codigos/madr/manage.py runserver'
test = '''
pytest -s -x --cov=madr --cov-report=html -vv && xdg-open htmlcov/index.html
'''
